requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in --output-file=requirements.txt --python=3.11
altair==6.0.0
    # via streamlit
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.12.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
attrs==25.4.0
    # via
    #   jsonschema
    #   referencing
av==16.0.1
    # via faster-whisper
blinker==1.9.0
    # via streamlit
boto3==1.42.12
    # via -r requirements.in
botocore==1.42.12
    # via
    #   boto3
    #   s3transfer
cachetools==6.2.4
    # via streamlit
certifi==2025.11.12
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via
    #   streamlit
    #   typer-slim
    #   uvicorn
coloredlogs==15.0.1
    # via onnxruntime
ctranslate2==4.6.2
    # via faster-whisper
diskcache==5.6.3
    # via llama-cpp-python
fastapi==0.125.0
    # via -r requirements.in
faster-whisper==1.2.1
    # via -r requirements.in
filelock==3.20.1
    # via huggingface-hub
flatbuffers==25.9.23
    # via onnxruntime
fsspec==2025.12.0
    # via huggingface-hub
gitdb==4.0.12
    # via gitpython
gitpython==3.1.45
    # via streamlit
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
hf-xet==1.2.0
    # via huggingface-hub
httpcore==1.0.9
    # via httpx
httptools==0.7.1
    # via uvicorn
httpx==0.28.1
    # via huggingface-hub
huggingface-hub==1.2.3
    # via
    #   -r requirements.in
    #   faster-whisper
    #   tokenizers
humanfriendly==10.0
    # via coloredlogs
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
jinja2==3.1.6
    # via
    #   altair
    #   llama-cpp-python
    #   pydeck
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsonschema==4.25.1
    # via altair
jsonschema-specifications==2025.9.1
    # via jsonschema
llama-cpp-python==0.3.16
    # via -r requirements.in
markupsafe==3.0.3
    # via jinja2
mpmath==1.3.0
    # via sympy
narwhals==2.14.0
    # via altair
numpy==2.3.5
    # via
    #   ctranslate2
    #   llama-cpp-python
    #   onnxruntime
    #   pandas
    #   pydeck
    #   streamlit
onnxruntime==1.23.2
    # via faster-whisper
packaging==25.0
    # via
    #   altair
    #   huggingface-hub
    #   onnxruntime
    #   streamlit
pandas==2.3.3
    # via streamlit
pillow==12.0.0
    # via streamlit
protobuf==6.33.2
    # via
    #   onnxruntime
    #   streamlit
pyarrow==22.0.0
    # via streamlit
pydantic==2.12.5
    # via fastapi
pydantic-core==2.41.5
    # via pydantic
pydeck==0.9.1
    # via streamlit
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   pandas
python-dotenv==1.2.1
    # via uvicorn
python-multipart==0.0.21
    # via -r requirements.in
pytz==2025.2
    # via pandas
pyyaml==6.0.3
    # via
    #   ctranslate2
    #   huggingface-hub
    #   uvicorn
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.5
    # via streamlit
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
s3transfer==0.16.0
    # via boto3
setuptools==80.9.0
    # via ctranslate2
shellingham==1.5.4
    # via huggingface-hub
six==1.17.0
    # via python-dateutil
smmap==5.0.2
    # via gitdb
starlette==0.50.0
    # via fastapi
streamlit==1.52.2
    # via -r requirements.in
sympy==1.14.0
    # via onnxruntime
tenacity==9.1.2
    # via streamlit
tokenizers==0.22.1
    # via faster-whisper
toml==0.10.2
    # via streamlit
tornado==6.5.4
    # via streamlit
tqdm==4.67.1
    # via
    #   faster-whisper
    #   huggingface-hub
typer-slim==0.20.0
    # via huggingface-hub
typing-extensions==4.15.0
    # via
    #   altair
    #   anyio
    #   fastapi
    #   huggingface-hub
    #   llama-cpp-python
    #   pydantic
    #   pydantic-core
    #   referencing
    #   starlette
    #   streamlit
    #   typer-slim
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.3
    # via pandas
urllib3==2.6.2
    # via
    #   botocore
    #   requests
uvicorn==0.38.0
    # via -r requirements.in
uvloop==0.22.1
    # via uvicorn
watchdog==6.0.0
    # via streamlit
watchfiles==1.1.1
    # via uvicorn
websockets==15.0.1
    # via uvicorn
